@@ -1,40 +1,80 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net7.0-windows</TargetFramework>
+    <TargetFramework>net8.0-windows</TargetFramework>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <Configurations>Debug;Release;Personal</Configurations>
   </PropertyGroup>
 
   <ItemGroup>
     <Compile Include="Elmish.Extensions.fs" />
-    <Compile Include="Abstractions.fs" />
     <Compile Include="Types.fs" />
+    <Compile Include="Abstractions.fs" />
     <Compile Include="Infrastructure.fs" />
+    <Compile Include="LoggerExtensions.fs" />
+    <Compile Include="WorkEventDetailsModel\WorkEventDetailsModel.fs" />
+    <Compile Include="WorkEventModel\WorkEventModel.fs" />
+    <Compile Include="WorkEventListModel\WorkEventListModel.fs" />
+    <Compile Include="WorkModel\WorkModel.fs" />
+    <Compile Include="WorkModel\WorkModel.Program.fs" />
+    <Compile Include="WorkModel\WorkModel.Bindings.fs" />
+    <Compile Include="AddWorkTimeModel\AddWorkTimeModel.fs" />
+    <Compile Include="WorkStatisticModel\WorkStatisticModel.fs" />
+    <Compile Include="WorkStatisticModel\WorkStatisticModel.Program.fs" />
+    <Compile Include="WorkStatisticModel\WorkStatisticModel.Bindings.fs" />
+    <Compile Include="WorkStatisticListModel\WorkStatisticListModel.fs" />
+    <Compile Include="WorkStatisticListModel\WorkStatisticListModel.Program.fs" />
+    <Compile Include="WorkStatisticListModel\WorkStatisticListModel.Bindings.fs" />
+    <Compile Include="CreatingWorkModel\CreatingWorkModel.fs" />
+    <Compile Include="CreatingWorkModel\CreatingWorkModel.Program.fs" />
+    <Compile Include="CreatingWorkModel\CreatingWorkModel.Bindings.fs" />
+    <Compile Include="WorkListModel\WorkListModel.fs" />
+    <Compile Include="WorkListModel\WorkListModel.Program.fs" />
+    <Compile Include="WorkListModel\WorkListModel.Bindings.fs" />
+    <Compile Include="WorkSelectorModel\WorkSelectorModel.fs" />
+    <Compile Include="WorkSelectorModel\WorkSelectorModel.Program.fs" />
+    <Compile Include="WorkSelectorModel\WorkSelectorModel.Bindings.fs" />
     <Compile Include="BotSettingsModel\BotSettingsModel.fs" />
     <Compile Include="BotSettingsModel\BotSettingsModel.Program.fs" />
     <Compile Include="BotSettingsModel\BotSettingsModel.Bindings.fs" />
+    <Compile Include="RollbackWorkModel\RollbackWorkModel.fs" />
+    <Compile Include="TimePointPrototypeModel\TimePointPrototypeModel.fs" />
+    <Compile Include="TimePointModel\TimePointModel.fs" />
+    <Compile Include="TimePointsGeneratorModel\TimePointsGeneratorModel.fs" />
+    <Compile Include="TimePointsGeneratorModel\TimePointsGeneratorModel.Program.fs" />
+    <Compile Include="TimePointsGeneratorModel\TimePointsGeneratorModel.Bindings.fs" />
+    <Compile Include="AppDialogModel\AppDialogModel.fs" />
+    <Compile Include="AppDialogModel\AppDialogModel.Program.fs" />
+    <Compile Include="AppDialogModel\AppDialogModel.Bindings.fs" />
     <Compile Include="MainModel\MainModel.fs" />
     <Compile Include="MainModel\MainModel.Program.fs" />
     <Compile Include="MainModel\MainModel.Bindings.fs" />
+    <Compile Include="CompositionRoot.fs" />
     <Compile Include="Program.fs" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Elmish.WPF" Version="4.0.0-beta-45" />
+    <PackageReference Include="Elmish.WPF" Version="4.0.0-beta-54" />
+    <PackageReference Include="FSharp.SystemTextJson" Version="1.3.13" />
+    <PackageReference Include="FsToolkit.ErrorHandling" Version="4.15.1" />
     <PackageReference Include="Ookii.Dialogs.Wpf" Version="5.0.1" />
-    <PackageReference Include="Serilog.Extensions.Logging" Version="3.1.0" />
-    <PackageReference Include="Serilog.Sinks.Console" Version="4.1.0" />
+    <PackageReference Include="Serilog.Extensions.Logging" Version="8.0.0" />
+    <PackageReference Include="Serilog.Sinks.Console" Version="5.0.1" />
     <PackageReference Include="Serilog.Sinks.Debug" Version="2.0.0" />
-    <PackageReference Include="Telegram.Bot" Version="18.0.0" />
+    <PackageReference Include="Telegram.Bot" Version="19.0.0" />
   </ItemGroup>
 
   <ItemGroup>
     <InternalsVisibleTo Include="PomodoroWindowsTimer.WpfClient" />
+    <InternalsVisibleTo Include="PomodoroWindowsTimer.ElmishApp.Tests" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\PomodoroWindowsTimer\PomodoroWindowsTimer.fsproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Update="FSharp.Core" Version="8.0.300-beta.24178.1" />
   </ItemGroup>
 
 </Project>
