@@ -13,37 +13,15 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace PomodoroWindowsTimer.WpfClient
+namespace PomodoroWindowsTimer.WpfClient;
+
+/// <summary>
+/// Interaction logic for MainWindow.xaml
+/// </summary>
+public partial class MainWindow : Window
 {
-    /// <summary>
-    /// Interaction logic for MainWindow.xaml
-    /// </summary>
-    public partial class MainWindow : Window
+    public MainWindow()
     {
-        public MainWindow()
-        {
-            InitializeComponent();
-        }
-
-        private void Slider_MouseDown(object sender, MouseButtonEventArgs e)
-        {
-            if (e.LeftButton == MouseButtonState.Pressed)
-            {
-                ICommand openFileCommand = ((dynamic)DataContext).PreChangeActiveTimeSpanCommand;
-                if (openFileCommand.CanExecute(null))
-                {
-                    openFileCommand.Execute(null);
-                }
-            }
-        }
-
-        private void Slider_MouseUp(object sender, MouseButtonEventArgs e)
-        {
-            ICommand openFileCommand = ((dynamic)DataContext).PostChangeActiveTimeSpanCommand;
-            if (openFileCommand.CanExecute(null))
-            {
-                openFileCommand.Execute(null);
-            }
-        }
+        InitializeComponent();
     }
 }
