@@ -5,14 +5,18 @@ open FsUnit
 open System.Diagnostics
 open NUnit.Framework.Constraints
 open System.Threading.Tasks
+open FParsec
+
+
+[<RequireQualifiedAccess>]
+module Async =
+    let asTask computation : Task = Async.StartAsTask computation :> _
 
 [<AutoOpen>]
 module FsUnit =
 
     let inline writeLine s = TestContext.WriteLine(sprintf "%A" s)
     let inline writeLineS (s: string) = TestContext.WriteLine(s)
-
-    let toTask computation : Task = Async.StartAsTask computation :> _
 
 
     [<DebuggerNonUserCode>]
@@ -29,6 +33,29 @@ module FsUnit =
             else
                 let divider = String.replicate 40 "-"
                 Assert.That(y, c, fun _ -> sprintf "%s\n  %s" message divider)
+
+
+[<AutoOpen>]
+module FParsec =
+
+    [<DebuggerStepThrough>]
+    let shouldSuccessEqual res = function
+    | Success (s, _, _) -> s |> FsUnit.shouldL equal res ""
+    | Failure (t, _, _) -> raise (AssertionException($"Should be %A{res} but was %A{t}"))
+
+
+    [<DebuggerStepThrough>]
+    let shouldSuccess ``constraint`` res = function
+    | Success (s, _, _) -> s |> should ``constraint`` res
+    | Failure (t, _, _) -> raise (AssertionException($"Should be %A{res} but was %A{t}"))
+
+
+    [<DebuggerStepThrough>]
+    let runResult p input =
+        runParserOnString p () "test" input
+        |> function
+            | Success (ok,_,_) -> Result.Ok ok
+            | Failure (err,_,_) -> Result.Error err
 
 
 [<RequireQualifiedAccess>]
